select precision

+++ resblock.py
@@ -3,12 +3,12 @@ import torch.nn as nn
 
 
 class ResBlock(nn.Module):
-    def __init__(self, num_features):
+    def __init__(self, num_features, dtype=torch.float32):
         super().__init__()
-        self.conv1 = nn.Conv2d(num_features, num_features, 3, 1, 1)
-        self.conv2 = nn.Conv2d(num_features, num_features, 3, 1, 1)
-        self.bnorm1 = nn.BatchNorm2d(num_features)
-        self.bnorm2 = nn.BatchNorm2d(num_features)
+        self.conv1 = nn.Conv2d(num_features, num_features, 3, 1, 1, dtype=dtype)
+        self.conv2 = nn.Conv2d(num_features, num_features, 3, 1, 1, dtype=dtype)
+        self.bnorm1 = nn.BatchNorm2d(num_features, dtype=dtype)
+        self.bnorm2 = nn.BatchNorm2d(num_features, dtype=dtype)
         self.nonlinear = nn.ReLU(inplace=True)
         nn.init.xavier_uniform_(self.conv1.weight)
         nn.init.xavier_uniform_(self.conv2.weight)
